# feat: optimize skiplist_mask with O(1) lookup table approach