@NohTow 2025-09-08
Hey,
Thanks for the PR!

From my understand, the main speed-up should be thanks to apply masking through the indexing rather than the loop (sorry about the loop btw, should have coded it more cleanly in the first place, was in a rush for the first release back then).
Is the LUT really beneficial? I feel like the skiplist is fixed, at least for a given instantiated model so we could store only a single list (as we are already doing with words/ids right now) and use it for indexing

Am I missing something?